refactor(inline,textnode): move `split_nodes_delimiter` function to inline.py
